requirements.txt
# Core Python ML stack
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0

# Gradient Boosting Models
xgboost>=2.0.0
lightgbm>=4.2.0

# Model persistence
joblib>=1.3.0

# Backend API
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0

# Data acquisition (Kaggle)
kaggle>=1.6.12

# Optional but production-useful
python-dotenv>=1.0.1
